Static towers, level loading
Towers can be set as static which disables the ability to remove them or modify their key. There is now a level loading function in main that removes / instantiates level scenes in a new container. The root node also now has a list of scenes that it treats as individual levels.